server/requirements.txt
django-cors-headers==3.14.0
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
pytz==2022.7.1
gunicorn==20.1.0
psycopg2-binary==2.9.1
drf_yasg


# epiz_34158315